all products endpoint
